Break up controller
